+++ client/src/Specification/pages/SpecificationList/index.tsx
@@ -1,11 +1,15 @@
 import React, { useEffect, useState } from 'react'
 import axios, { AxiosResponse } from 'axios'
-import { Table } from 'antd'
-import { useRouteMatch } from 'react-router-dom'
 import { useHistory } from 'react-router-dom'
 import PageTitle from 'src/shared/components/PageTitle'
-import TableHeader from '../../components/TableHeader'
 import SpecificationFormModal from '../../components/SpecificationFormModal'
+import SpecificationTable from '../../components/SpecificationTable'
+
+export interface ISpecification {
+  id: string
+  name: string
+  updated_at: string
+}
 
 const SpecificationList: React.SFC = () => {
   const [specifications, setSpecifications] : any = useState([])
@@ -14,27 +18,7 @@ const SpecificationList: React.SFC = () => {
   const [editId, setEditId] = useState('')
 
   const [visible, setVisible] = useState(false)
-  let match = useRouteMatch()
   let history = useHistory()
-
-  const onCreateClick = () => {
-    setModalInitialValue({name: ''})
-    setModalType('create')
-    setEditId('')
-    setVisible(true)
-  }
-
-  const onEditClick = (id: string) => {
-    const getSpecificationItem = async () => {
-      const response: AxiosResponse = await axios.get(`http://localhost:3000/api/v1/specifications/${id}`)
-
-      setModalInitialValue({ name: response.data.name })
-      setModalType('edit')
-      setEditId(id)
-      setVisible(true)
-    }
-    getSpecificationItem()
-  }
 
   useEffect(() => {
     const getApiResult = async () => {
@@ -43,79 +27,6 @@ const SpecificationList: React.SFC = () => {
     }
     getApiResult()
   }, [])
-
-  interface ISpecification {
-    id: string
-    name: string
-    updated_at: string
-  }
-
-  const dataSource = Object.keys(specifications).length === 0 ? [] :
-  specifications.map((specification: ISpecification) => (
-      {
-        key: specification.id,
-        name: specification.name,
-        updated_at: specification.updated_at,
-      }
-    ))
-
-  const deleteSpecificationItem = async (id: string) => {
-    await axios.delete(`http://localhost:3000/api/v1/specifications/${id}`)
-    const nextSpecifications: ISpecification[] = specifications.filter((specification: ISpecification) => ( specification.id !== id ))
-    setSpecifications(nextSpecifications)
-    history.push(`/specifications/`)
-  }
-
-  const specificationItemColumn = 
-    [
-      {
-        title: '仕様書名',
-        dataIndex: 'name',
-        key: 'name',
-      },
-      {
-        title: '更新日',
-        dataIndex: 'updated_at',
-        key: 'updated_at',
-      },
-      {
-        title: '',
-        key: 'action',
-        render: (record: any) => {
-          return (
-            <button onClick={() => {
-              history.push(`/specifications/${record.key}/specification_items`)
-            }}>
-              仕様書項目
-            </button>
-          )
-        },
-      },
-      {
-        title: '',
-        key: 'action',
-        render: (record: any) => {
-          return (
-            <button onClick={() => onEditClick(record.key)}>
-              編集
-            </button>
-          )
-        },
-      },
-      {
-        title: '',
-        key: 'action',
-        render: (record: any) => {
-          return (
-            <button onClick={() => {
-              deleteSpecificationItem(record.key) 
-            }}>
-              削除
-            </button>
-          )
-        },
-      },
-    ]
 
   const onCreate = async (values: { name: string }) => {
     try {
@@ -126,8 +37,7 @@ const SpecificationList: React.SFC = () => {
         id: result.data.id, name: result.data.name, updated_at: result.data.updated_at
       }])
       setVisible(false)
-      console.log(specifications)
-      setSpecifications(...specifications, nextSpecifications)
+      setSpecifications(nextSpecifications)
     } catch(error) {
       history.push(`/specifications/new`)
     }
@@ -140,7 +50,6 @@ const SpecificationList: React.SFC = () => {
       })
       const updatedSpacification = { id: result.data.id, name: result.data.name, updated_at: result.data.updated_at }
       const nextSpecifications: ISpecification[] = specifications.map((specification: ISpecification) => {
-        debugger
         return specification.id === updatedSpacification.id ? updatedSpacification : specification
       })
       setVisible(false)
@@ -153,11 +62,18 @@ const SpecificationList: React.SFC = () => {
   const onCancel = () => {
     setVisible(false);
   }
+
   return (
     <>
       <PageTitle>仕様書一覧</PageTitle>
-      <TableHeader onCreateClick={onCreateClick} />
-      <Table dataSource={dataSource} columns={specificationItemColumn}/>
+      <SpecificationTable
+        specifications={specifications}
+        setModalInitialValue={setModalInitialValue}
+        setModalType={setModalType}
+        setEditId={setEditId}
+        setVisible={setVisible}
+        setSpecifications={setSpecifications}
+      />
       <SpecificationFormModal
         visible={visible}
         onCreate={onCreate}
